LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/document-rules.https_include=selectorMatchesDisplayNone-expected.txt


FAIL test selector_matches with link inside display:none container assert_equals: expected 0 but got 1
